Moved logic around to make cpan find freshly installed modules
Immediatly after installing a module, CPAN could not find it.

I moved the @INC initialization uppward, included the
archname-version also. We now update both @INC and set the PERL5LIB.

We delay the loading of CPAN until this @INC is setup, but I was getting
errors with subclassing CPAN::Prompt because of this, so we loose the
Local::Prompt.

Signed-off-by: Pedro Melo <melo@simplicidade.org>